# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AlternativePayload": {
        "properties": {
          "id": {
            "description": "Caller's identifier for this alternative",
            "title": "Id",
            "type": "string"
          },
          "levels": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Level keys for indicators declared on an ordered or unordered scale.",
            "title": "Levels",
            "type": "object"
          },
          "values": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Numeric indicator values. Omit an indicator you have no value for; omission means unknown and is never imputed.",
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id"
        ],
        "title": "AlternativePayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScoreRequest": {
        "properties": {
          "alternatives": {
            "description": "The shortlist to rank.",
            "items": {
              "$ref": "#/components/schemas/AlternativePayload"
            },
            "maxItems": 5,
            "title": "Alternatives",
            "type": "array"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "context": {
            "description": "Active application contexts. Defaults to the category's baseline context when empty.",
            "items": {
              "type": "string"
            },
            "title": "Context",
            "type": "array"
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          }
        },
        "required": [
          "category",
          "alternatives"
        ],
        "title": "ScoreRequest",
        "type": "object"
      },
      "ScoreResponse": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "contexts": {
            "items": {
              "type": "string"
            },
            "title": "Contexts",
            "type": "array"
          },
          "eligibility_precondition": {
            "title": "Eligibility Precondition",
            "type": "string"
          },
          "functional_unit": {
            "title": "Functional Unit",
            "type": "string"
          },
          "model_snapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Snapshot"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "registry_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Registry Version"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/ScoredAlternative"
            },
            "title": "Results",
            "type": "array"
          },
          "stakeholders": {
            "items": {
              "type": "string"
            },
            "title": "Stakeholders",
            "type": "array"
          }
        },
        "required": [
          "category",
          "functional_unit",
          "eligibility_precondition",
          "contexts",
          "stakeholders",
          "registry_version",
          "model_snapshot",
          "results",
          "notes"
        ],
        "title": "ScoreResponse",
        "type": "object"
      },
      "ScoredAlternative": {
        "properties": {
          "disqualifying_levels": {
            "items": {
              "type": "string"
            },
            "title": "Disqualifying Levels",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "missing_indicators": {
            "items": {
              "type": "string"
            },
            "title": "Missing Indicators",
            "type": "array"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "id",
          "score",
          "rank",
          "missing_indicators",
          "disqualifying_levels"
        ],
        "title": "ScoredAlternative",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Inference API.\n\nSubmit a shortlist and a context, receive a score per alternative. The response\nalways echoes the category's declared eligibility precondition, because the\nmodel ranks options that are assumed to have already passed regulatory\nprefiltering and saying so is the difference between a recommendation and a\ncompliance claim.\n\nThe registry travels inside the served model file, so the served semantics are\nexactly the ones the weights were trained under, with no database involved.\n",
    "title": "Context-adaptive product recommender",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/categories/{category_key}/indicators": {
      "get": {
        "operationId": "indicators",
        "parameters": [
          {
            "in": "path",
            "name": "category_key",
            "required": true,
            "schema": {
              "title": "Category Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Indicators",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Indicators"
      }
    },
...
    "/api/explore/response": {
      "get": {
        "operationId": "response",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "required": true,
            "schema": {
              "title": "Category",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "indicator",
            "required": true,
            "schema": {
              "title": "Indicator",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "context",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Context"
            }
          },
          {
            "in": "query",
            "name": "stakeholders",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stakeholders"
            }
          },
          {
            "in": "query",
            "name": "steps",
            "required": false,
            "schema": {
              "default": 21,
              "title": "Steps",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Response",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Response",
        "tags": [
          "explore"
        ]
      }
    },
...
    "/api/score": {
      "post": {
        "operationId": "score",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Score"
      }
    },
...
  }
}